# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MyPay fintech API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upi:
    get:
      operationId: getUpiAccounts
      tags: [upi]
      summary: List user UPI accounts
      responses:
        "200":
          description: UPI accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiAccountsResult"
    post:
      operationId: addUpiAccount
      tags: [upi]
      summary: Add a UPI account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpiAccountInput"
      responses:
        "201":
          description: Account added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiAccountCreated"
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Browse available tasks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksResult"
...
  /tasks/submissions:
    get:
      operationId: getMyTaskSubmissions
      tags: [tasks]
      summary: Get own task submissions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSubmissionsResult"

  # ── REFERRALS ────────────────────────────────────────────────
...
  /admin/tasks:
    get:
      operationId: adminListTasks
      tags: [admin]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTasksResult"
    post:
      operationId: adminCreateTask
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCreated"
  /admin/tasks/{id}:
    patch:
      operationId: adminUpdateTask
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCreated"
    delete:
      operationId: adminDeleteTask
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/tasks/submissions:
    get:
      operationId: adminListSubmissions
      tags: [admin]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSubmissionsResult"
  /admin/tasks/submissions/{id}/approve:
    patch:
      operationId: adminApproveSubmission
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/tasks/submissions/{id}/reject:
    patch:
      operationId: adminRejectSubmission
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── ADMIN UPI ────────────────────────────────────────────────
  /admin/upi:
    get:
      operationId: adminListUpiAccounts
      tags: [admin]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: UPI accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUpiResult"
  /admin/upi/{id}/approve:
    patch:
      operationId: adminApproveUpi
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNotesInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/upi/{id}/reject:
    patch:
      operationId: adminRejectUpi
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── ADMIN SETTINGS ───────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string

    # ── Auth schemas ──────────────────────────────────────────
...
    UpiAccountItem:
      type: object
      properties:
        id:
          type: string
        upiId:
          type: string
        accountHolderName:
          type: string
        provider:
          type: string
        status:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
    UpiAccountsResult:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/UpiAccountItem"
    UpiAccountInput:
      type: object
      required: [upiId, accountHolderName, provider]
      properties:
        upiId:
          type: string
        accountHolderName:
          type: string
        provider:
          type: string
    UpiAccountCreated:
      type: object
      properties:
        success:
          type: boolean
        account:
          $ref: "#/components/schemas/UpiAccountItem"

    # ── Tasks ────────────────────────────────────────────────
    TaskItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        amount:
          type: string
        durationDays:
          type: integer
        status:
          type: string
        imageUrl:
          type: ["string", "null"]
        totalSubmissions:
          type: integer
        maxSubmissions:
          type: ["integer", "null"]
        createdAt:
          type: string
    TasksResult:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskItem"
        hasMore:
          type: boolean
...
    TaskSubmissionItem:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        proofImageUrl:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        rejectedReason:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskSubmissionsResult:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/TaskSubmissionItem"
        hasMore:
          type: boolean

    # ── Referrals ────────────────────────────────────────────
...
    ReasonInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    AdminNotesInput:
      type: object
      properties:
        adminNotes:
          type: string

    # ── Admin Deposits ───────────────────────────────────────
...
    TaskInput:
      type: object
      required: [name, description, amount]
      properties:
        name:
          type: string
        description:
          type: string
        amount:
          type: number
        durationDays:
          type: integer
        maxSubmissions:
          type: integer
        imageUrl:
          type: string
    TaskUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        amount:
          type: number
        isActive:
          type: boolean
        imageUrl:
          type: string
    TaskCreated:
      type: object
      properties:
        success:
          type: boolean
        task:
          $ref: "#/components/schemas/TaskItem"
    AdminTasksResult:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskItem"
        hasMore:
          type: boolean
    AdminSubmissionItem:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        userId:
          type: string
        proofImageUrl:
          type: string
        status:
          type: string
        taskName:
          type: ["string", "null"]
        taskAmount:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userMobile:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminSubmissionsResult:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/AdminSubmissionItem"
        hasMore:
          type: boolean

    # ── Admin UPI ────────────────────────────────────────────
    AdminUpiItem:
      type: object
      properties:
        id:
          type: string
        upiId:
          type: string
        accountHolderName:
          type: string
        provider:
          type: string
        status:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        userMobile:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminUpiResult:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AdminUpiItem"
        hasMore:
          type: boolean

    # ── Admin Settings ───────────────────────────────────────
...